+++ teamfeedback/proxy-server/public/index.html
@@ -7,26 +7,26 @@
   <body>
     <script crossorigin src="https://unpkg.com/react@16/umd/react.development.js"></script>
     <script crossorigin src="https://unpkg.com/react-dom@16/umd/react-dom.development.js"></script>
-    <!-- <div id="navbar"></div> -->
+    <div id="navbar"></div>
     <div id="descriptions"></div>
     <div id="Tiers"></div>
     <div id="individual-items"></div>
-    <!-- <div id="reviews"></div> -->
+    <div id="reviews"></div>
     <div id="bonuses"></div>
-    <!-- <div id="contributors"></div> -->
-    <!-- <div id="charity"></div> -->
-    <!-- <div id="purchase"></div> -->
-    <!-- <div id="system-requirements"></div> -->
+    <div id="contributors"></div>
+    <div id="charity"></div>
+    <div id="purchase"></div>
+    <div id="system-requirements"></div>
     <!-- <script src="http://localhost:3102"></script> Navbar (secondary) -->
-    <script src="http://localhost:3663/bundle.js"></script> <!-- Descriptions (primary) -->
-    <script src="http://localhost:3101/static/js/bundle.js"></script> <!-- Tiers (primary) -->
-    <script src="http://localhost:3200/bundle.js"></script> <!-- Individual Items (primary) -->
+    <!-- <script src="http://localhost:3663/bundle.js"></script> Descriptions (primary) -->
+    <!-- <script src="http://localhost:3101/static/js/bundle.js"></script> Tiers (primary) -->
+    <!-- <script src="http://localhost:3200/bundle.js"></script> Individual Items (primary) -->
     <!-- <script src="http://localhost:3030"></script> Reviews (secondary) -->
-    <script src="http://localhost:3031/bundle.js"></script> <!-- Bonuses (primary) -->
+    <!-- <script src="http://localhost:3031/bundle.js"></script> Bonuses (primary) -->
     <!-- <script src="http://localhost:3876"></script> Contributors (secondary) -->
     <!-- <script src="http://localhost:3987" ></script> Charity (primary) -->
     <!-- <script src="http://localhost:3201"></script> Purchase (secondary) -->
-    <!-- <script src="http://localhost:3201"></script> System Requirements (secondary) -->
+    <script src="http://localhost:3201/dist/bundle/bundle.js"></script> System Requirements (secondary)
   </body>
   <footer></footer>
 </html>
