--- teamfeedback/reverse-proxy-server/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "proxy",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "server": "nodemon server.js"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "express": "^4.17.1",
-    "nodemon": "^2.0.4"
-  },
-  "dependencies": {
-    "axios": "^0.20.0"
-  }
-}
